requirements.txt
Flask
pymongo
python-dotenv
dnspython
faiss-cpu
huggingface_hub
torch
gunicorn
google-generativeai
sentence-transformers
transformers
tokenizers
selenium
beautifulsoup4
requests
--extra-index-url https://download.pytorch.org/whl/cpu